high-nb fix in chiral model EoS module
Sometimes in the simulations, the ratio of baryon density to energy
density hits the unpysical region. In such case, chiral model EoS module
returns zero pressure. When such zero-pressure zone emerges in the core
of the fluid blob, it results in energy density clumping.

We assume that the unphysically high baryon density is accidental,
therefore a rather simple fix is applied to the Chiral model EoS module:
if the evaluated pressure is very low, the EoS is re-evaluated at
a somewhat lower baryon density (and the same energy density).